+++ src/fst_impls/vector/vector_fst.rs
@@ -3,8 +3,9 @@ use arc::Arc;
 use failure::{bail, ensure, format_err};
 use fst_traits::{
     ArcIterator, CoreFst, ExpandedFst, FinalStatesIterator, Fst, MutableArcIterator, MutableFst,
-    StateIterator,
+    StateIterator, TextParser,
 };
+use parsers::text::ParsedTextFst;
 use semirings::Semiring;
 use std::fmt;
 use std::ops::{Add, BitOr};
@@ -14,8 +15,14 @@ use StateId;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct VectorFst<W: Semiring> {
-    states: Vec<VectorFstState<W>>,
-    start_state: Option<StateId>,
+    pub(crate) states: Vec<VectorFstState<W>>,
+    pub(crate) start_state: Option<StateId>,
+}
+
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct VectorFstState<W: Semiring> {
+    pub(crate) final_weight: Option<W>,
+    pub(crate) arcs: Vec<Arc<W>>,
 }
 
 impl<W: 'static + Semiring> Fst for VectorFst<W> {}
@@ -182,15 +189,41 @@ impl<'a, W: 'static + Semiring> MutableArcIterator<'a> for VectorFst<W> {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct VectorFstState<W: Semiring> {
-    final_weight: Option<W>,
-    arcs: Vec<Arc<W>>,
-}
-
 impl<W: Semiring> VectorFstState<W> {
     pub fn num_arcs(&self) -> usize {
         self.arcs.len()
+    }
+}
+
+impl<W: 'static + Semiring<Type = f32>> TextParser for VectorFst<W> {
+    fn from_parsed_fst_text(parsed_fst_text: ParsedTextFst) -> Result<Self> {
+        let start_state = parsed_fst_text.start();
+        let num_states = parsed_fst_text.num_states();
+
+        let states = vec![VectorFstState::<W>::default(); num_states];
+
+        let mut fst = VectorFst {
+            states,
+            start_state,
+        };
+
+        for transition in parsed_fst_text.transitions.into_iter() {
+            let weight = transition.weight.map(W::new).unwrap_or_else(W::one);
+            let arc = Arc::new(
+                transition.ilabel,
+                transition.olabel,
+                weight,
+                transition.nextstate,
+            );
+            fst.add_arc(&transition.state, arc)?;
+        }
+
+        for final_state in parsed_fst_text.final_states.into_iter() {
+            let weight = final_state.weight.map(W::new).unwrap_or_else(W::one);
+            fst.set_final(&final_state.state, weight)?;
+        }
+
+        Ok(fst)
     }
 }