+++ src/semirings/integer_weight.rs
@@ -9,14 +9,12 @@ pub struct IntegerWeight {
     value: i32,
 }
 
-impl IntegerWeight {
-    pub fn new(value: i32) -> Self {
-        IntegerWeight { value }
-    }
-}
-
 impl Semiring for IntegerWeight {
     type Type = i32;
+
+    fn new(value: <Self as Semiring>::Type) -> Self {
+        IntegerWeight { value }
+    }
 
     fn plus(&self, rhs: &Self) -> Self {
         Self::new(self.value + rhs.value)
@@ -52,6 +50,6 @@ impl StarSemiring for IntegerWeight {
         if self.value == 0 {
             return Self::new(1);
         }
-        return Self::new(i32::max_value());
+        Self::new(i32::max_value())
     }
 }