+++ src/test_data/text_fst/macros.rs
@@ -1,0 +1,5 @@
+macro_rules! rel_to_abs_path {
+    ($filename: expr) => {
+        PathBuf::from(file!()).parent().unwrap().join($filename)
+    };
+}